+++ src/shared/lib/constants.ts
@@ -1,0 +1,10 @@
+// Centralized constants for the Explore module
+
+export const filterTypes = [
+  'Type',
+  'Skill level',
+  'Group type',
+  'Duration',
+  'Distance',
+  'Credit Range',
+]
