+++ src/styles/themes.css
@@ -22,7 +22,6 @@
   --error: #d32f2f;
   --info: #308be4;
   --icon-input: #828282;
-  --badge-color: #faddd5;
 
   --background: oklch(1 0 0); /* #ffffff */
   --foreground: oklch(0 0 0); /* #000000 */
